Bulk delete on control panel page

+++ includes/page-poll-dude-control-panel.php
@@ -32,28 +32,33 @@ if(!empty($_POST['do'])) {
 if (isset($_POST['bulk_delete'])) {
     global $poll_dude, $wpdb;
     //check_ajax_referer('wp-polls_bulk-delete');
+    echo $_POST['delete_all']."\n";
     check_admin_referer( 'wp-polls_bulk-delete' );
-    for($i=0; $i<count($_POST['pollq']); $i++){
-        $pollq_id = $_POST['pollq'][$i];
-        
-        $pollq_question = $wpdb->get_var( $wpdb->prepare( "SELECT pollq_question FROM $wpdb->pollsq WHERE pollq_id = %d", $pollq_id ) );
-        $poll_question_text = wp_kses_post( $poll_dude->utility->removeslashes($pollq_question));        
-        $delete_poll_question = $wpdb->delete( $wpdb->pollsq, array( 'pollq_id' => $pollq_id ), array( '%d' ) );
-        $delete_poll_answers =  $wpdb->delete( $wpdb->pollsa, array( 'polla_qid' => $pollq_id ), array( '%d' ) );
-        $delete_poll_ip =	   $wpdb->delete( $wpdb->pollsip, array( 'pollip_qid' => $pollq_id ), array( '%d' ) );
-        $poll_option_lastestpoll = $wpdb->get_var("SELECT option_value FROM $wpdb->options WHERE option_name = 'poll_latestpoll'");        
-        
-        $error = false;
-        if(!$delete_poll_question) {
-            echo '<p style="color: red;">'.sprintf(__('Error In Deleting Poll \'%s\' Question', 'poll-dude-domain'), wp_kses_post( $poll_dude->utility->removeslashes( $pollq_question ) ) ).'</p>';
-            $error = true;
+    if(isset($_POST['delete_all'])){
+        echo '<p style="color: green;">'.sprintf(__('Poll Deleted all', 'poll-dude-domain')).'</p>';
+    } else {
+        for($i=0; $i<count($_POST['pollq']); $i++){
+            $pollq_id = $_POST['pollq'][$i];
+            
+            $pollq_question = $wpdb->get_var( $wpdb->prepare( "SELECT pollq_question FROM $wpdb->pollsq WHERE pollq_id = %d", $pollq_id ) );
+            $poll_question_text = wp_kses_post( $poll_dude->utility->removeslashes($pollq_question));        
+            $delete_poll_question = $wpdb->delete( $wpdb->pollsq, array( 'pollq_id' => $pollq_id ), array( '%d' ) );
+            $delete_poll_answers =  $wpdb->delete( $wpdb->pollsa, array( 'polla_qid' => $pollq_id ), array( '%d' ) );
+            $delete_poll_ip =	   $wpdb->delete( $wpdb->pollsip, array( 'pollip_qid' => $pollq_id ), array( '%d' ) );
+            $poll_option_lastestpoll = $wpdb->get_var("SELECT option_value FROM $wpdb->options WHERE option_name = 'poll_latestpoll'");        
+            
+            $error = false;
+            if(!$delete_poll_question) {
+                echo '<p style="color: red;">'.sprintf(__('Error In Deleting Poll \'%s\' Question', 'poll-dude-domain'), wp_kses_post( $poll_dude->utility->removeslashes( $pollq_question ) ) ).'</p>';
+                $error = true;
+            }
+            if(!$error) {
+                echo '<p style="color: green;">'.sprintf(__('Poll \'%d\' \'%s\' Deleted Successfully', 'poll-dude-domain'), $pollq_id, wp_kses_post( $poll_dude->utility->removeslashes( $pollq_question ) ) ).'</p>';
+            }
+                    
+            update_option( 'poll_latestpoll', $poll_dude->utility->latest_poll() );
+            //do_action( 'wp_polls_delete_poll', $pollq_id );
         }
-        if(!$error) {
-            echo '<p style="color: green;">'.sprintf(__('Poll \'%d\' \'%s\' Deleted Successfully', 'poll-dude-domain'), $pollq_id, wp_kses_post( $poll_dude->utility->removeslashes( $pollq_question ) ) ).'</p>';
-        }
-                
-        update_option( 'poll_latestpoll', $poll_dude->utility->latest_poll() );
-        //do_action( 'wp_polls_delete_poll', $pollq_id );
     }
 }
 
@@ -101,10 +106,9 @@ switch($mode) {
                 <thead>
                     <tr>
                         <th></th>
-                        <th><input id="cb-select-all1" type="checkbox" name="bulk_delete" value="all" style="width:10px; height:15px;" ></th>
+                        <th><input id="cb-select-all1" type="checkbox" name="delete_all" value="delete_all" style="width:10px; height:15px;" ></th>
                         <th></th>
                         <th colspan="7"><?php 
-                            //wp_create_nonce('wp-polls_bulk-delete');
                             wp_nonce_field( 'wp-polls_bulk-delete' );
                             echo "<input class=\"button-secondary\" name=\"bulk_delete\" type=\"submit\" value=\"".__('Bulk Delete', 'poll-dude-domain')." \" />\n";
                         ?></th>
@@ -202,10 +206,9 @@ switch($mode) {
                     ?>
                     <tr>
                         <th></th>
-                        <th><input id="cb-select-all2" type="checkbox" name="bulk_delete" value="all" style="width:10px; height:15px;" ></th>
+                        <th><input id="cb-select-all2" type="checkbox" name="delete_all" value="delete_all" style="width:10px; height:15px;" ></th>
                         <th></th>
                         <th><?php 
-                            //wp_create_nonce('wp-polls_bulk-delete');
                             wp_nonce_field( 'wp-polls_bulk-delete' );
                             echo "<input class=\"button-secondary\" name=\"bulk_delete\" type=\"submit\" value=\"".__('Bulk Delete', 'poll-dude-domain')." \" />\n";
                         ?></th>
